requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.21.0
gymnasium>=0.28.0
pyyaml>=6.0
matplotlib>=3.5.0
tensorboard>=2.8.0

# MATE environment dependencies
pygame>=2.1.0
pyglet>=1.5.0
scipy>=1.7.0

# Optional dependencies for enhanced functionality
wandb>=0.12.0  # For experiment tracking
tqdm>=4.62.0   # For progress bars
seaborn>=0.11.0  # For better plotting
jupyter>=1.0.0  # For notebooks

# Development dependencies
pytest>=6.2.0
black>=21.0.0
flake8>=4.0.0
mypy>=0.910
